pyproject.toml
[project]
name = "sesame-wake"
version = "0.1.0"
description = "Sesame voice launcher: bring-your-own openWakeWord ONNX + Chrome"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
dependencies = [
  "openwakeword",
  "pyaudio",
  "numpy",
  "selenium",
  "python-dotenv",
  "textual>=8.2.6",
]

[project.urls]
Homepage = "https://github.com/emasuriano/sesame-wake"
Repository = "https://github.com/emasuriano/sesame-wake"
Issues = "https://github.com/emasuriano/sesame-wake/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["sesame_wake"]

[project.scripts]
sesame-wake = "sesame_wake.cli:main"

[dependency-groups]
dev = [
  "ruff>=0.15.12",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = [
  "E",   # pycodestyle errors
  "W",   # pycodestyle warnings
  "F",   # pyflakes
  "I",   # isort
  "UP",  # pyupgrade
  "B",   # flake8-bugbear
  "SIM", # flake8-simplify
]
